Adding a CGI which triggers build_admin_war.sh so it can be called via HTTP from frinex_restart_experient.cgi

+++ src/main/cgi/frinex_restart_experient.cgi
@@ -38,6 +38,10 @@ if [ -f /FrinexBuildService/protected/$experimentDirectory/$cleanedInput.war ]; 
             echo "$version is >= 1.8.0"
         else
             echo "$version is < 1.8.0"
+            echo "this war needs to be updated, renaming old war to $(cleanedInput}_${version}.war"
+            mv /FrinexBuildService/protected/$experimentDirectory/$cleanedInput.war /FrinexBuildService/protected/$experimentDirectory/$(cleanedInput}_${version}.war
+            echo "Please reload this page<br>"
+            echo "<button onClick=\"window.location.reload();\">Refresh Page</button>"
         fi
     fi
     if  [ "$QUERY_STRING" == "$cleanedInput&actuator/health" ] || [ "$QUERY_STRING" == "$cleanedInput&health" ]; then
@@ -106,8 +110,8 @@ if [ -f /FrinexBuildService/protected/$experimentDirectory/$cleanedInput.war ]; 
             echo "<button onClick=\"window.location.reload();\">Refresh Page</button>"
         fi
     fi
-else
-    echo "The experiment $cleanedInput does not exist."
+else if [ -f /FrinexBuildService/artifacts/$experimentDirectory/$experimentDirectory.xml ]; then
+    echo "The experiment ${cleanedInput}.war does not exist."
     echo "$(date), not found, $cleanedInput, $QUERY_STRING" >> /usr/local/apache2/htdocs/frinex_restart_experient.log
     if [[ "$cleanedInput" == *_staging_admin || "$cleanedInput" == *_production_admin ]]; then
         echo "Building admin war for $cleanedInput<br>"
@@ -115,4 +119,6 @@ else
         echo "Please reload this page in a few minutes<br>"
         echo "<button onClick=\"window.location.reload();\">Refresh Page</button>"
     fi
+else
+    echo "The experiment ${experimentDirectory}.xml does not exist."    
 fi
